requirements.txt
azure-ai-projects
azure-ai-inference
azure-ai-agents
azure-cosmos
azure-search-documents
azure-identity
python-dotenv
fastapi[all]
uvicorn[standard]
httpx
slowapi>=0.1.9
email-validator